+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 ## 0.4.3 (2020-03-26)
+Changes:
+- Updated spark-cassandra-connector from 2.4.2 to 2.5.0
+- Updated spark-excel-connector from 0.12.4 to 0.13.1
+- Updated spark-dynamodb-connector from 1.0.1 to 1.0.4
+- Updated scalatest (scope test) from 3.1.0 to 3.1.2
+- Updated postgresql (scope test) from 42.2.9 to 42.2.12
+
 New Features:
 - Added pipeline dependency check before starting the spark job (#113)
 - Added default Spark job group and description (#115)
+
+Fixes
 
 ## 0.4.2 (2020-02-15)
 - Fixed cross building issue (#111)